+++ fracture/src/fracture/utils/logging.h
@@ -12,6 +12,15 @@
 #include "fracture/fracture_core.h"
 
 /**
+ * @brief The source of the log message.
+ * 
+ */
+typedef enum log_source {
+    LOG_SOURCE_CORE = 0,
+    LOG_SOURCE_CLIENT = 1
+} log_source;
+
+/**
  * @brief Flags to enable or disable logging levels at bit position corresponding to the log level.
  * 
  */
@@ -68,9 +77,6 @@ typedef struct logging_config {
     b8 logging_flags;
 } logging_config;
 
-// HACK: This is just to test platform layer
-FR_API void application_loop();
-
 /**
  * @brief Initialize the logging system with the given configuration. The amount of memory required for the logging system is returned in the memory_required parameter if the config pointer passed is null.
  *        It is the responsibility of the caller to allocate the required memory and pass it to the function.
@@ -79,13 +85,13 @@ FR_API void application_loop();
  * @param memory_required The amount of memory required for the logging system.
  * @return b8 True if the logging system was successfully initialized, false otherwise.
  */
-FR_API b8 initialize_logging(logging_config* config);
+b8 initialize_logging(logging_config* config);
 
 /**
  * @brief Shutdown the logging system, closing any open files and freeing any allocated memory. 
  * 
  */
-FR_API void shutdown_logging();
+void shutdown_logging();
 
 /**
  * @brief Log a message with the given level, file, line, and format string. 
@@ -96,7 +102,7 @@ FR_API void shutdown_logging();
  * @param format The format string for the log message.
  * @param ... The arguments to the format string.
  */
-FR_API void log_message_detailed(log_level level, const char* file, int line, const char* format, ...);
+FR_API void log_message_detailed(log_level level, log_source source, const char* file, int line, const char* format, ...);
 
 /**
  * @brief Log a message with the given level and format string. 
@@ -105,7 +111,7 @@ FR_API void log_message_detailed(log_level level, const char* file, int line, co
  * @param format The format string for the log message.
  * @param ... The arguments to the format string.
  */
-FR_API void log_message(log_level level, const char* format, ...);
+FR_API void log_message(log_level level, log_source source, const char* format, ...);
 
 /**
  * @brief Check if the given logging level is enabled. 
@@ -113,7 +119,7 @@ FR_API void log_message(log_level level, const char* format, ...);
  * @param level The logging level to check.
  * @return b9 True if the logging level is enabled, false otherwise.
  */
-b8 logging_level_get(log_level level);
+FR_API b8 logging_level_get(log_level level);
 
 /**
  * @brief Set the logging level to the given value. 
@@ -156,35 +162,35 @@ FR_API void logging_file_set(b8 enabled);
  * 
  * @param filename The filename to write the log to.
  */
-void logging_filename_set(const char* filename);
+FR_API void logging_filename_set(const char* filename);
 
 // Define logging macros
 
-#define FR_CORE_FATAL(format, ...) log_message(LOG_LEVEL_FATAL, format, ##__VA_ARGS__)
-#define FR_CORE_FATAL_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_FATAL, __FILE__, __LINE__, format, ##__VA_ARGS__)
+#define FR_CORE_FATAL(format, ...) log_message(LOG_LEVEL_FATAL, LOG_SOURCE_CORE, format, ##__VA_ARGS__)
+#define FR_CORE_FATAL_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_FATAL, LOG_SOURCE_CORE, __FILE__, __LINE__, format, ##__VA_ARGS__)
 
-#define FR_CORE_ERROR(format, ...) log_message(LOG_LEVEL_ERROR, format, ##__VA_ARGS__)
-#define FR_CORE_ERROR_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_ERROR, __FILE__, __LINE__, format, ##__VA_ARGS__)
+#define FR_CORE_ERROR(format, ...) log_message(LOG_LEVEL_ERROR, LOG_SOURCE_CORE, format, ##__VA_ARGS__)
+#define FR_CORE_ERROR_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_ERROR, LOG_SOURCE_CORE, __FILE__, __LINE__, format, ##__VA_ARGS__)
 
 #if ENABLE_WARN_LOGGING == 1
-    #define FR_CORE_WARN(format, ...) log_message(LOG_LEVEL_WARN, format, ##__VA_ARGS__)
-    #define FR_CORE_WARN_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_WARN, __FILE__, __LINE__, format, ##__VA_ARGS__)
+    #define FR_CORE_WARN(format, ...) log_message(LOG_LEVEL_WARN, LOG_SOURCE_CORE, format, ##__VA_ARGS__)
+    #define FR_CORE_WARN_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_WARN, LOG_SOURCE_CORE, __FILE__, __LINE__, format, ##__VA_ARGS__)
 #else
     #define FR_CORE_WARN(format, ...)
     #define FR_CORE_WARN_DETAILED(format, ...)
 #endif
 
 #if ENABLE_INFO_LOGGING == 1
-    #define FR_CORE_INFO(format, ...) log_message(LOG_LEVEL_INFO, format, ##__VA_ARGS__)
-    #define FR_CORE_INFO_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_INFO, __FILE__, __LINE__, format, ##__VA_ARGS__)
+    #define FR_CORE_INFO(format, ...) log_message(LOG_LEVEL_INFO, LOG_SOURCE_CORE, format, ##__VA_ARGS__)
+    #define FR_CORE_INFO_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_INFO, LOG_SOURCE_CORE, __FILE__, __LINE__, format, ##__VA_ARGS__)
 #else
     #define FR_CORE_INFO(format, ...)
     #define FR_CORE_INFO_DETAILED(format, ...)
 #endif
 
 #if ENABLE_TRACE_LOGGING == 1
-    #define FR_CORE_TRACE(format, ...) log_message(LOG_LEVEL_TRACE, format, ##__VA_ARGS__)
-    #define FR_CORE_TRACE_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_TRACE, __FILE__, __LINE__, format, ##__VA_ARGS__)
+    #define FR_CORE_TRACE(format, ...) log_message(LOG_LEVEL_TRACE, LOG_SOURCE_CORE, format, ##__VA_ARGS__)
+    #define FR_CORE_TRACE_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_TRACE, LOG_SOURCE_CORE, __FILE__, __LINE__, format, ##__VA_ARGS__)
 #else
     #define FR_CORE_TRACE(format, ...)
     #define FR_CORE_TRACE_DETAILED(format, ...)
@@ -193,7 +199,46 @@ void logging_filename_set(const char* filename);
 // Assert macros
 
 #if FR_ENABLE_ASSERTS
-    #define FR_CORE_ASSERT(expression) if (!(expression)) { log_message_detailed(LOG_LEVEL_ASSERT, __FILE__, __LINE__, #expression); DEBUG_BREAK(); }
+    #define FR_CORE_ASSERT(expression) if (!(expression)) { log_message_detailed(LOG_LEVEL_ASSERT, LOG_SOURCE_CORE, __FILE__, __LINE__, #expression); DEBUG_BREAK(); }
 #else
     #define FR_CORE_ASSERT(expression)
+#endif
+
+
+// Client logging macros
+
+#define FR_FATAL(format, ...) log_message(LOG_LEVEL_FATAL, LOG_SOURCE_CLIENT, format, ##__VA_ARGS__)
+#define FR_FATAL_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_FATAL, LOG_SOURCE_CLIENT, __FILE__, __LINE__, format, ##__VA_ARGS__)
+
+#define FR_ERROR(format, ...) log_message(LOG_LEVEL_ERROR, LOG_SOURCE_CLIENT, format, ##__VA_ARGS__)
+#define FR_ERROR_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_ERROR, LOG_SOURCE_CLIENT, __FILE__, __LINE__, format, ##__VA_ARGS__)
+
+#if ENABLE_WARN_LOGGING == 1
+    #define FR_WARN(format, ...) log_message(LOG_LEVEL_WARN, LOG_SOURCE_CLIENT, format, ##__VA_ARGS__)
+    #define FR_WARN_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_WARN, LOG_SOURCE_CLIENT, __FILE__, __LINE__, format, ##__VA_ARGS__)
+#else
+    #define FR_WARN(format, ...)
+    #define FR_WARN_DETAILED(format, ...)
+#endif
+
+#if ENABLE_INFO_LOGGING == 1
+    #define FR_INFO(format, ...) log_message(LOG_LEVEL_INFO, LOG_SOURCE_CLIENT, format, ##__VA_ARGS__)
+    #define FR_INFO_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_INFO, LOG_SOURCE_CLIENT, __FILE__, __LINE__, format, ##__VA_ARGS__)
+#else
+    #define FR_INFO(format, ...)
+    #define FR_INFO_DETAILED(format, ...)
+#endif
+
+#if ENABLE_TRACE_LOGGING == 1
+    #define FR_TRACE(format, ...) log_message(LOG_LEVEL_TRACE, LOG_SOURCE_CLIENT, format, ##__VA_ARGS__)
+    #define FR_TRACE_DETAILED(format, ...) log_message_detailed(LOG_LEVEL_TRACE, LOG_SOURCE_CLIENT, __FILE__, __LINE__, format, ##__VA_ARGS__)
+#else
+    #define FR_TRACE(format, ...)
+    #define FR_TRACE_DETAILED(format, ...)
+#endif
+
+#if FR_ENABLE_ASSERTS
+    #define FR_ASSERT(expression) if (!(expression)) { log_message_detailed(LOG_LEVEL_ASSERT, LOG_SOURCE_CLIENT, __FILE__, __LINE__, #expression); DEBUG_BREAK(); }
+#else
+    #define FR_ASSERT(expression)
 #endif